refactor : refactor alarm
