+++ app/src/main/java/com/nbcproject/earlybirdy/util/SharedPreferenceKeys.kt
@@ -1,0 +1,10 @@
+package com.nbcproject.earlybirdy.util
+
+class SharedPreferenceKeys {
+    companion object {
+        const val alarmSetting = "alarmSetting"
+        const val hour = "hour"
+        const val minute = "minute"
+        const val alarmSwitch = "alarmSwitch"
+    }
+}
